+++ src/main/java/me/despical/kotl/ConfigPreferences.java
@@ -54,7 +54,7 @@ public class ConfigPreferences {
 
 	public enum Option {
 		BOSSBAR_ENABLED("Bossbar-Enabled", true), CHAT_FORMAT_ENABLED("ChatFormat-Enabled", true),
-		CLEAR_EFFECTS("Clear-Effects", true), CLEAR_INVENTORY("Clear-Inventory", true),
+		CLEAR_EFFECTS("Clear-Effects", true), CLEAR_INVENTORY("Clear-Inventory", false),
 		DATABASE_ENABLED("DatabaseActivated", false), DEATHBLOCKS_ENABLED("Death-Blocks.Enabled", true),
 		DISABLE_FALL_DAMAGE("Disable-Fall-Damage", true), INVENTORY_MANAGER_ENABLED("InventoryManager", true),
 		JOIN_NOTIFY("Join-Notify", true), LEAVE_NOTIFY("Leave-Notify", true), SCOREBOARD_ENABLED("Scoreboard-Enabled", false),